requirements.txt
requests==2.31.0
python-telegram-bot==20.3
feedparser==6.0.10
schedule==1.2.0
python-dotenv==1.0.0
flask==3.0.0
flask-login==0.6.3
werkzeug==3.0.1
gunicorn==21.2.0
